Skip over nil funcs
Without this, if a user passes a nil func to us, we'll panic with a
really confusing nil pointer dereference later on, when context on the
original cause has been lost.

The alternative to this is to panic with a helpful message at the call
site, to let the caller know they are misusing the library. The problem
is we don't know where we're being called from, and a panic may be
really inappropriate.

A further improvement might be to have an optional logging mechanism to
emit warnings about this, or an option to panic in these cases.